Use explicit casts in all multiplications that might overflow.
This is primarily relevant for pointer arithmetic where a multiplication
is used to compute an offset in a dense 2D array; if the dimensions are
stored as smaller integer types, the product may overflow that type if
it is not explicitly forced to be size_t.

+++ include/tatami/other/DelayedBind.hpp
@@ -609,7 +609,7 @@ public:
 
         double denom = 0;
         for (const auto& x : mats) {
-            double total = x->nrow() * x->ncol();
+            double total = static_cast<double>(x->nrow()) * static_cast<double>(x->ncol());
             denom += total;
             sparse_prop += total * x->sparse_proportion();
             row_prop += total * x->prefer_rows_proportion();
